refactor: Rename segmentation settings component
